# contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: ProcureIQ Sales API
  description: API for managing CRM Sales Orders, Products, Accounts, and Notifications in ProcureIQ.
  version: 1.2.0
paths:
...
  /api/v1/fieldservice/territories:
    post:
      summary: Create service territory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTerritoryRequest'
      responses:
        '201':
          description: Service territory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceTerritoryResponse'
  /api/v1/fieldservice/territories/{id}:
    get:
      summary: Retrieve service territory details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service territory retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceTerritoryResponse'
    put:
      summary: Update service territory details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTerritoryRequest'
      responses:
        '200':
          description: Service territory updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceTerritoryResponse'
    delete:
      summary: Soft-delete service territory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service territory deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/v1/fieldservice/work-types:
    post:
      summary: Create work type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkTypeRequest'
      responses:
        '201':
          description: Work type created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkTypeResponse'
  /api/v1/fieldservice/work-types/{id}:
    get:
      summary: Retrieve work type details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Work type retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkTypeResponse'
    put:
      summary: Update work type details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkTypeRequest'
      responses:
        '200':
          description: Work type updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkTypeResponse'
    delete:
      summary: Soft-delete work type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Work type deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/v1/fieldservice/skills:
    post:
      summary: Create skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillRequest'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSkillResponse'
  /api/v1/fieldservice/skills/{id}:
    get:
      summary: Retrieve skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSkillResponse'
    put:
      summary: Update skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSkillResponse'
    delete:
      summary: Delete skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
...
    ApiResponseString:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        data:
          type: string
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    ServiceTerritoryRequest:
      type: object
      required:
        - name
        - operatingHoursId
      properties:
        name:
          type: string
        operatingHoursId:
          type: integer
        isActive:
          type: boolean
    ServiceTerritoryResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        operatingHoursId:
          type: integer
        isActive:
          type: boolean
    ApiResponseServiceTerritoryResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/ServiceTerritoryResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    WorkTypeRequest:
      type: object
      required:
        - name
        - estimatedDuration
        - durationType
      properties:
        name:
          type: string
        estimatedDuration:
          type: number
        durationType:
          type: string
          enum: [hours, minutes]
        shouldAutoCreateAppointment:
          type: boolean
    WorkTypeResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        estimatedDuration:
          type: number
        durationType:
          type: string
        shouldAutoCreateAppointment:
          type: boolean
    ApiResponseWorkTypeResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/WorkTypeResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    SkillRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    SkillResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    ApiResponseSkillResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/SkillResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...